refactoring css
