+++ src/_scss/pages/homepageUpdate/homepageUpdate.scss
@@ -102,6 +102,7 @@
   @import "../homepage/download/download";
   @import "../homepage/community/community";
   @import "hero/hero";
+  @import "hero/rotatingText";
   @import "./homepageCovidSection";
 }
