--- src/js/components/homepageUpdate/hero/HeroPoc.scss
@@ -1,246 +1,0 @@
-.rotation-text-container {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    margin-bottom: 4rem;
-    @media (max-width: 992px) {
-        display: block;
-        margin: 0;
-    }
-}
-
-.hero__heading-wrapper {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    @media (max-width: 468px) {
-        display: block !important;
-        padding: 0 !important;
-    }
-}
-
-.sentence {
-    display: flex;
-    flex-wrap: wrap;
-    justify-content: center;
-    @media (max-width: 992px) {
-        display: block;
-        width: 640px;
-    }
-    @media (max-width: 468px) {
-        display: block;
-        width: unset;
-    }
-}
-
-.landingPhrase {
-    position: absolute;
-    @media (max-width: 992px) {
-        display: block;
-        text-align: left;
-        max-width: 400px;
-    }
-}
-
-.phraseIntro {
-    width: 300px;
-    flex-shrink: 0;
-    white-space: nowrap;
-
-    @media (max-width: 992px) {
-        display: block;
-        //position: relative;
-        width: unset;
-        flex-shrink: unset;
-        height: 50px;
-        text-align: left;
-        left: 0;
-    }
-}
-
-.phraseEnd {
-    width: 400px;
-    flex-shrink: 0;
-    white-space: nowrap;
-    height: 70px;
-    @media (max-width: 992px) {
-        display: block;
-    }
-}
-
-.staticBlock {
-    white-space: nowrap;
-    flex-shrink: 0;
-    color: #1b1b1b;
-    //flex-basis: 540px;
-    @media (max-width: 992px) {
-        display: block;
-        text-align: left;
-        white-space: break-spaces !important;
-    }
-}
-
-/*Vertical Sliding*/
-.leftWords {
-    position: relative;
-    span {
-        position: absolute;
-        right: 0;
-        @media (max-width: 992px) {
-            left: 0;
-        }
-    }
-
-    @media (max-width: 992px) {
-        left: 0;
-    }
-}
-
-.rightWords {
-    display: flex;
-    flex-wrap: wrap;
-    -webkit-align-self: center;
-    justify-content: center;
-    justify-self: center;
-    align-content: center;
-    align-self: center;
-    position: relative;
-    span {
-        position: absolute;
-        left: 0;
-    }
-    @media (min-width: 1200px) {
-        display: flex;
-        padding-top: 0;
-        position: relative;
-        white-space: nowrap;
-        span {
-            position: absolute;
-            left: 0;
-        }
-    }
-}
-
-.initialVertical span {
-    animation: centerToBottom 20s linear;
-    -ms-animation: centerToBottom 20s linear;
-    -webkit-animation: centerToBottom 20s linear;
-    color: #2d6878;
-    opacity: 0;
-    overflow: hidden;
-}
-
-.finalVertical span {
-    animation: fadeElIn 25s;
-    -ms-animation: fadeElIn 25s;
-    -webkit-animation: fadeElIn 25s;
-    color: #2d6878;
-    opacity: 1;
-    overflow: hidden;
-}
-
-.wholePhrase {
-    -webkit-animation: fadeElOut 20s;
-    //-webkit-animation-delay: 20s;
-    overflow: hidden;
-    opacity: 0;
-}
-
-/* fade out */
-//@-moz-keyframes topToCenter{
-//    0% { opacity: 0; -moz-transform: translateY(-50px); }
-//    90% { opacity: 0; -moz-transform: translateY(-50px); }
-//    95% { opacity: 1; -moz-transform: translateY(0px); }
-//    100% { opacity: 1; -moz-transform: translateY(0px); }
-//
-//}
-
-@-webkit-keyframes fadeElIn{
-    0% { opacity: 0 }
-    75% { opacity: 0 }
-    80% { opacity: 1 }
-    100% { opacity: 1 }
-
-}
-
-@-webkit-keyframes fadeElOut{
-    0% { opacity: 1 }
-    90% { opacity: 1 }
-    100% { opacity: 0 }
-
-}
-
-.slidingVertical span {
-    animation: topToBottom 25s linear;
-    -ms-animation: topToBottom 25s linear;
-    -webkit-animation: topToBottom 25s linear;
-    color: #2d6878;
-    opacity: 0;
-    overflow: hidden;
-    //position: absolute;
-}
-
-.slidingVertical span:nth-child(2){
-    animation-delay: 5s;
-    -ms-animation-delay: 5s;
-    -webkit-animation-delay: 5s;
-}
-.slidingVertical span:nth-child(3){
-    animation-delay: 10s;
-    -ms-animation-delay: 10s;
-    -webkit-animation-delay: 10s;
-}
-.slidingVertical span:nth-child(4){
-    animation-delay: 15s;
-    -ms-animation-delay: 15s;
-    -webkit-animation-delay: 15s;
-}
-.slidingVertical span:nth-child(5){
-    animation-delay: 20s;
-    -ms-animation-delay: 20s;
-    -webkit-animation-delay: 20s;
-}
-/* fade out */
-@-moz-keyframes centerToBottom{
-    0% { opacity: 1; -moz-transform: translateY(0px); }
-    5% { opacity: 1; -moz-transform: translateY(0px); }
-    10% { opacity: 0; -moz-transform: translateY(50px); }
-    100% { opacity: 0; -moz-transform: translateY(50px); }
-
-}
-
-@-webkit-keyframes centerToBottom{
-    0% { opacity: 1; -webkit-transform: translateY(0px); }
-    5% { opacity: 1; -webkit-transform: translateY(0px); }
-    10% { opacity: 0; -webkit-transform: translateY(50px); }
-    100% { opacity: 0; -webkit-transform: translateY(50px); }
-}
-
-/*topToBottom Animation*/
-@-moz-keyframes topToBottom{
-    0% { opacity: 0; }
-    5% { opacity: 0; -moz-transform: translateY(-50px); }
-    10% { opacity: 1; -moz-transform: translateY(0px); }
-    25% { opacity: 1; -moz-transform: translateY(0px); }
-    30% { opacity: 0; -moz-transform: translateY(50px); }
-    80% { opacity: 0; }
-    100% { opacity: 0; }
-}
-@-webkit-keyframes topToBottom{
-    0% { opacity: 0; }
-    5% { opacity: 0; -webkit-transform: translateY(-50px); }
-    10% { opacity: 1; -webkit-transform: translateY(0px); }
-    25% { opacity: 1; -webkit-transform: translateY(0px); }
-    30% { opacity: 0; -webkit-transform: translateY(50px); }
-    80% { opacity: 0; }
-    100% { opacity: 0; }
-}
-@-ms-keyframes topToBottom{
-    0% { opacity: 0; }
-    5% { opacity: 0; -ms-transform: translateY(-50px); }
-    10% { opacity: 1; -ms-transform: translateY(0px); }
-    25% { opacity: 1; -ms-transform: translateY(0px); }
-    30% { opacity: 0; -ms-transform: translateY(50px); }
-    80% { opacity: 0; }
-    100% { opacity: 0; }
-}
