+++ src/js/registerIcons.js
@@ -32,7 +32,6 @@ import {
     faDownload,
     faEllipsisH,
     faEnvelope,
-    faEquals,
     faExclamationCircle,
     faExternalLinkAlt,
     faExclamationTriangle,
@@ -51,7 +50,9 @@ import {
     faMinus,
     faMinusSquare,
     faMoneyCheckAlt,
+    faPause,
     faPencilAlt,
+    faPlay,
     faPlus,
     faPlusCircle,
     faSearch,
@@ -96,7 +97,6 @@ export default function registerIcons() {
         faCubes,
         faDownload,
         faEllipsisH,
-        faEquals,
         faEnvelope,
         faExclamationCircle,
         faExternalLinkAlt,
@@ -117,7 +117,9 @@ export default function registerIcons() {
         faMinus,
         faMinusSquare,
         faMoneyCheckAlt,
+        faPause,
         faPencilAlt,
+        faPlay,
         faPlus,
         faPlusCircle,
         faSearch,
